+++ NEWS.md
@@ -1,5 +1,7 @@
 # bumbl 1.0.2
 
+-   Added a new dataset, `test_df`, containing simulated data used for package testing.
+
 # bumbl 1.0.1
 
 -   Updated package citation and author list